Vietnt

+++ routes/admin.php
@@ -58,8 +58,8 @@ Route::prefix('admin')
             // CATALOGUES
             Route::resource('catalogues', CatalogueController::class);
             //ACTIVATE
-            Route::post('catalogues/{catalogue}/activate', [CatalogueController::class, 'activate'])->name('catalogues.activate');
-            Route::post('catalogues/{catalogue}/deactivate', [CatalogueController::class, 'deactivate'])->name('catalogues.deactivate');
+            Route::post('catalogues/{catalogue}/activate', [CatalogueController::class, 'activate'])->name('catalogues.activate')->middleware('permission:Kích hoạt danh mục');
+            Route::post('catalogues/{catalogue}/deactivate', [CatalogueController::class, 'deactivate'])->name('catalogues.deactivate')->middleware('permission:Kích hoạt danh mục');
 
             // PRODUCT
             Route::resource('products', ProductController::class);
@@ -70,33 +70,34 @@ Route::prefix('admin')
             Route::get('/variants/{id}/edit', [ProductVariantController::class, 'edit']);
             Route::put('/variants/{id}', [ProductVariantController::class, 'update']);
             Route::delete('/variants/{id}', [ProductVariantController::class, 'destroy'])->name('variants.destroy');
-            Route::post('/products/variant/update', [ProductController::class, 'updateVariant'])->name('products.variant.update'); //chinh sú hang loat
-            Route::patch('/products/{id}/restore', [ProductController::class, 'restore'])->name('products.restore');
+            Route::post('/products/variant/update', [ProductController::class, 'updateVariant'])->name('products.variant.update')->middleware('permission:Chỉnh sửa sản phẩm'); //chinh sú hang loat
+            Route::patch('/products/{id}/restore', [ProductController::class, 'restore'])->name('products.restore')->middleware('permission:Khôi phục sản phẩm');
 
             // ORDER
             Route::resource('orders', OrderController::class);
             Route::get('orders/{id}', [OrderController::class, 'show'])->name('order.show');
-            Route::put('orders/update/{id}', [OrderController::class, 'update'])->name('order.update');
+            Route::put('orders/update/{id}', [OrderController::class, 'update'])->name('order.update')->middleware('permission:Chỉnh sửa trạng thái đơn hàng');
 
             // USER
             Route::resource('users', UserController::class);
-            Route::post('/user/{user}/active', [UserController::class, 'active'])->name('users.active');
-            Route::get('clients', [UserController::class, 'getAllClient'])->name('users.clients');
-            Route::get('staffs', [UserController::class, 'getAllStaff'])->name('users.staffs');
-            Route::get('staffs/create', [UserController::class, 'createStaff'])->name('users.staffs.create');
-            Route::post('staffs/store', [UserController::class, 'storeStaff'])->name('users.staffs.store');
-            Route::get('staffs/{user}/edit', [UserController::class, 'editStaff'])->name('users.staffs.edit');
-            Route::put('staffs/{user}/update', [UserController::class, 'updateStaff'])->name('users.staffs.update');
-            Route::post('/user/{user}/active', [UserController::class, 'active'])->name('users.active');
+            // Route::post('/user/{user}/active', [UserController::class, 'active'])->name('users.active');
+            Route::get('clients', [UserController::class, 'getAllClient'])->name('users.clients')->middleware('permission:xem danh sách khách hàng');
+            Route::get('staffs', [UserController::class, 'getAllStaff'])->name('users.staffs')->middleware('permission:xem danh sách nhân viên');
+            Route::get('staffs/create', [UserController::class, 'createStaff'])->name('users.staffs.create')->middleware('role:admin');
+            Route::post('staffs/store', [UserController::class, 'storeStaff'])->name('users.staffs.store')->middleware('role:admin');
+            Route::get('staffs/{user}/edit', [UserController::class, 'editStaff'])->name('users.staffs.edit')->middleware('role:admin');
+            Route::put('staffs/{user}/update', [UserController::class, 'updateStaff'])->name('users.staffs.update')->middleware('role:admin');
+            Route::put('staffs/{user}/quickly', [UserController::class, 'updateQuicklyRole'])->name('users.staffs.quickly')->middleware('role:admin');
+            Route::post('/user/{user}/active', [UserController::class, 'active'])->name('users.active')->middleware('role:admin');
 
             // Role
-            Route::get('roles', [UserController::class, 'getAllRole'])->name('users.roles');
-            Route::post('roles/store', [UserController::class, 'storeRole'])->name('users.roles.store');
-            Route::get('roles/{role}/edit', [UserController::class, 'editRole'])->name('users.roles.edit');
-            Route::get('roles/{role}/permission', [UserController::class, 'getAllPermissionRole'])->name('users.roles.permission');
-            Route::post('roles/{role}/permission/update', [UserController::class, 'updatePermissionRole'])->name('users.roles.permission.update');
-            Route::post('roles/{role}/update', [UserController::class, 'updateRole'])->name('users.roles.update');
-            Route::delete('roles/{role}/delete', [UserController::class, 'deleteRole'])->name('users.roles.delete');
+            Route::get('roles', [UserController::class, 'getAllRole'])->name('users.roles')->middleware('role:admin');
+            Route::post('roles/store', [UserController::class, 'storeRole'])->name('users.roles.store')->middleware('role:admin');
+            Route::get('roles/{role}/edit', [UserController::class, 'editRole'])->name('users.roles.edit')->middleware('role:admin');
+            Route::get('roles/{role}/permission', [UserController::class, 'getAllPermissionRole'])->name('users.roles.permission')->middleware('role:admin');
+            Route::post('roles/{role}/permission/update', [UserController::class, 'updatePermissionRole'])->name('users.roles.permission.update')->middleware('role:admin');
+            Route::post('roles/{role}/update', [UserController::class, 'updateRole'])->name('users.roles.update')->middleware('role:admin');
+            Route::delete('roles/{role}/delete', [UserController::class, 'deleteRole'])->name('users.roles.delete')->middleware('role:admin');
             // Route::view('/users/show', 'admin.users.show')->name('users.show');
 
             // profile
@@ -106,7 +107,7 @@ Route::prefix('admin')
 
             // BANNER
             Route::resource('banners', BannerController::class);
-            Route::post('banners/{banner}/activate', [BannerController::class, 'activate'])->name('banners.activate');
+            Route::post('banners/{banner}/activate', [BannerController::class, 'activate'])->name('banners.activate')->middleware('permission:Kích hoạt banner');
 
             // VOUCHER
             Route::resource('vouchers', VoucherController::class);
